Move loaders and parsers into `legacy` directory
