+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/index.js
@@ -3,10 +3,10 @@
 const path = require('path');
 const htmlparser = require('htmlparser2');
 
-const astUtils = require('../../utils/ast-utils');
-const ngUtils = require('../../utils/ng-utils');
+const astUtils = require('../../../utils/ast-utils');
+const ngUtils = require('../../../utils/ng-utils');
 const BaseParser = require('../BaseParser');
-const ParserResult = require('../../utils/parser-result');
+const ParserResult = require('../../../utils/parser-result');
 
 module.exports = class AngularDirectiveParser extends BaseParser {
   constructor() {
