+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/index.spec.js
