+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/injected.js
