+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/inline-concat.js
