+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/minimal.js
