+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/no-scope.js
