+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/split-definition-injected.js
