+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/template-url-function.js
