+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/two-directives.js
