+++ packages/hekla-core/src/legacy/parsers/AngularDirectiveParser/test-examples/webpack-loader-template.js
