+++ packages/hekla-core/src/legacy/parsers/AngularFactoryParser/index.spec.js
