+++ packages/hekla-core/src/legacy/parsers/AngularFactoryParser/test-examples/basic.js
