+++ packages/hekla-core/src/legacy/parsers/AngularFactoryParser/test-examples/split-definition.js
