+++ packages/hekla-core/src/legacy/parsers/DefaultParser/test-examples/commonjs.js
