+++ packages/hekla-core/src/legacy/parsers/DefaultParser/test-examples/es6.js
